fix: keep the CLI api bundle and MCP trust guard intact after the router thinning
The workflow terminal continuation moved into workspaceService, whose value
import dragged hookService -> ptc -> typescript into dist/cli/api.mjs and broke
its ESM load (ERR_AMBIGUOUS_MODULE_SYNTAX); it now lives in a standalone module
with only type-level workspaceService edges. The stale router-level MCP trust
wiring test is replaced by a table-driven testForApi trust-resolution test at
the service layer.
